+++ winit-sdk/src/main/java/io/github/ealenxie/winit/lma/Parcel.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * @author create by sch  2023/4/21 15:51
  * @version 1.0
@@ -15,25 +17,25 @@ public class Parcel {
      * 体积
      */
     @JsonProperty("volume")
-    private Double volume;
+    private BigDecimal volume;
     /**
      * 长，CM
      */
     @JsonProperty("length")
-    private Double length;
+    private BigDecimal length;
     /**
      * 宽
      */
     @JsonProperty("width")
-    private Double width;
+    private BigDecimal width;
     /**
      * 重量，KG
      */
     @JsonProperty("weight")
-    private Double weight;
+    private BigDecimal weight;
     /**
      * 高
      */
     @JsonProperty("height")
-    private Double height;
+    private BigDecimal height;
 }
