+++ winit-sdk/src/main/java/io/github/ealenxie/winit/oswh/PackageInfo.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * @author create by sch  2023/4/20 16:49
  * @version 1.0
@@ -15,25 +17,25 @@ public class PackageInfo {
      * 包裹长
      */
     @JsonProperty("length")
-    private Double length;
+    private BigDecimal length;
     /**
      * 包裹宽
      */
     @JsonProperty("width")
-    private Double width;
+    private BigDecimal width;
     /**
      * 包裹高
      */
     @JsonProperty("weight")
-    private Double weight;
+    private BigDecimal weight;
     /**
      * 包裹重
      */
     @JsonProperty("height")
-    private Double height;
+    private BigDecimal height;
     /**
      * 包裹体积
      */
     @JsonProperty("CBM")
-    private Double cbm;
+    private BigDecimal cbm;
 }
